requirements.txt
streamlit>=1.20
streamlit-extras>=0.1.0
streamlit-option-menu>=0.4.0
undetected-chromedriver>=3.2.4
selenium>=4.8.0
PyPDF2>=3.0.0
pdfplumber>=0.7.5
python-docx>=0.8.11
pytesseract>=0.3.10
Pillow>=9.0
google-genai>=0.10.0
pandas>=2.0
numpy>=1.25
pytest>=7.0
python-dotenv>=1.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.3  # optional: nếu không muốn cài, faiss import sẽ thất bại và code sẽ fallback
scikit-learn>=1.2
playwright>=1.40
prefect>=2.12
pyyaml>=6.0
prometheus-client>=0.15.0
sentry-sdk>=1.14.0
mlflow>=2.4.0
scipy>=1.11.0
